Cleanup and add stricter types
