+++ src/index.tsx
@@ -1,10 +1,10 @@
 import React from 'react';
 import ReactDOM from 'react-dom';
 import App from './App';
-import 'normalize.css';
 import { Provider } from 'react-redux';
 import configureStore from './configureStore';
 import * as serviceWorker from './serviceWorker';
+import 'normalize.css';
 
 const store = configureStore()
