+++ src/reducers/preferencesReducer.ts
@@ -1,13 +1,16 @@
-import { IPreferenceState, PreferenceTypes } from '../types';
+import { IPreferenceState, ActionTypes, UPDATE_PREFERENCE } from '../actions';
 
 const initialState: IPreferenceState = {
   skinTone: '',
 };
 
-function preferencesReducer(state = initialState, action: PreferenceTypes): IPreferenceState {
+function preferencesReducer(state = initialState, action: ActionTypes): IPreferenceState {
   switch(action.type) {
-    case 'UPDATE_PREFERENCE':
-      return state;
+    case UPDATE_PREFERENCE: {
+      return {
+        skinTone: state.skinTone,
+      };
+    }
     default:
       return state;
   }
